add app blocking api
- launches CarAppBlockingPolicyService and read Policy on startup
- also provides setAppBlockingPolicy API
- not working fully due to missng framework api but only sets internal
  state for now.
- also fixed some build warnings
- also renamed all permisisons to android.support.car
- remaining TODO:
  monitor app install / uninstall and re-activate policy
  plumb framework side policy setting

bug: 25261583

+++ libvehiclenetwork/libvehiclenetwork-audio-helper/src/VehicleNetworkAudioHelper.cpp
@@ -148,7 +148,7 @@ void VehicleNetworkAudioHelper::notifyStreamStopped(int32_t stream) {
 
 VehicleNetworkAudioHelper::StreamState& VehicleNetworkAudioHelper::getStreamStateLocked(
         int32_t streamNumber) {
-    if (streamNumber >= mStreamStates.size()) {
+    if (streamNumber >= (int32_t) mStreamStates.size()) {
         mStreamStates.insertAt(mStreamStates.size(), streamNumber - mStreamStates.size() + 1);
     }
     return mStreamStates.editItemAt(streamNumber);
